// helpers/types.ts
export enum CombatElement {
  PHYSICAL = 'PHYSICAL',
  ICE='ICE', 
  QUANTUM='QUANTUM',
  FIRE='FIRE',
  IMAGINARY='IMAGINARY',
  LIGHTNING='LIGHTNING',
  WIND='WIND'
}

export enum CharPath {
  NIHILITY = 'NIHILITY',
  ERUDITION = 'ERUDITION',
  HUNT = 'HUNT',
  PRESERVATION = 'PRESERVATION',
  DESTRUCTION = 'DESTRUCTION',
  ABUNDANCE = 'ABUNDANCE',
  HARMONY = 'HARMONY'
}